+++ app/models/User.go
@@ -9,7 +9,7 @@ type User struct {
 	TaxID     string `json:"taxid"`              // taxId
 	Password  string `json:"password,omitempty"` // รหัสผ่าน
 	Mobile    string `json:"mobile"`             // เบอร์โทร
-	RoleID    int16  `json:"role_id"`            // สิทะฺเข้าใช้
+	RoleID    uint   `json:"role_id"`            // สิทะฺเข้าใช้
 	Role      Role   `json:"role"`
 	OTP       string
 	OTPExpiry time.Time
